moving binning methods from Pad- to TelescopeAnalysis to use them for Pix as well
